Added TPC query plans.
Rewrote the SQL parser from scratch, because of the language features used in TPC that weren't used in graph and RDF SQL queries. The new plans are in SQLPlannerTest.scala. All of the old tests should still be passing.

Made some changes to the optimizer itself so that SQL queries would work:
    - Make sure we output a valid join tree from the GHD solver.
    - Some SQL features (e.g. selections that use OR to touch multiple attributes) mean some relations need to be in the same bag.

Added some extensions to the IR in Selection, Aggregation, and a new OrderBy object.